simpler index with separate invited info component
